Removed broken icon theme from install instruction
The icon theme linked in the install instructions is broken,
as well as not being a part of the gtk repository (#142) and shouldn't
be linked to on the website. If the icon theme is to be included
as part of the dracula project, it should be working and have an
official repository, to allow contribution and development to
occur.